[MAT-223] Modularise jshim. Commits include:
[MAT-223] Towards modularising jshim
[MAT-223] Attempt to fix OSX/Win linkage
[MAT-223] Type error in template.
[MAT-223] Pull ExprFactory out of jshim.
[MAT-223] Outline DispatchToXXXArrayOfArrays
[MAT-195] Generate call and caller graphs.
[MAT-223] debug print for debug build, change indents.
[MAT-223] Remove extraneous SUPPRESS_UNUSED
[MAT-223] Remove unnecessary headers
[MAT-223] centralise checking for debug build in macro.
[MAT-223] ExprFactory has no state so convert to functions.
