+++ include/exprfactory.hh
@@ -1,0 +1,16 @@
+/**
+ * Copyright (C) 2013 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+#include "expression.hh"
+#include "jni.h"
+
+using librdag::OGNumeric;
+
+namespace convert {
+
+OGNumeric* createExpression(jobject obj);
+
+} // namespace convert
